+++ manta-accounting/src/wallet/signer/mod.rs
@@ -38,7 +38,7 @@ use crate::{
     wallet::ledger::{self, Data},
 };
 use alloc::{boxed::Box, vec::Vec};
-use core::{cmp::max, convert::Infallible, fmt::Debug, hash::Hash};
+use core::{cmp::max, convert::Infallible, fmt::Debug, hash::Hash, ops::SubAssign};
 use manta_crypto::{
     accumulator::{
         Accumulator, BatchInsertion, ExactSizeAccumulator, FromItemsAndWitnesses, ItemHashFunction,
@@ -1640,7 +1640,10 @@ where
 
     /// Signs the `transaction`, generating transfer posts.
     #[inline]
-    pub fn sign(&mut self, transaction: Transaction<C>) -> Result<SignResponse<C>, SignError<C>> {
+    pub fn sign(&mut self, transaction: Transaction<C>) -> Result<SignResponse<C>, SignError<C>>
+    where
+        C::AssetValue: SubAssign,
+    {
         functions::sign(
             &self.parameters,
             self.state.accounts.as_ref(),
@@ -1664,6 +1667,7 @@ where
         request: ConsolidationPrerequest<C>,
     ) -> Result<SignResponse<C>, SignError<C>>
     where
+        C::AssetValue: SubAssign,
         C::Identifier: PartialEq,
     {
         functions::consolidate(
@@ -1741,6 +1745,7 @@ where
         transaction: Transaction<C>,
     ) -> Result<SignWithTransactionDataResponse<C>, SignError<C>>
     where
+        C::AssetValue: SubAssign,
         TransferPost<C>: Clone,
     {
         functions::sign_with_transaction_data(
@@ -1835,8 +1840,8 @@ where
                 max(self.state.assets.select(asset).values.len() - 1, 1)
             }
             Transaction::ToPublic(asset, _) => {
-                max(self.state.assets.select(asset).values.len() - 1, 1)
-            } // note: change the estimation once we implement the topublic optimization
+                (self.state.assets.select(asset).values.len() + 1) / 2
+            }
         }
     }
 }
@@ -1844,7 +1849,8 @@ where
 impl<C> Connection<C> for Signer<C>
 where
     C: Configuration,
-    C::AssetValue: CheckedAdd<Output = C::AssetValue> + CheckedSub<Output = C::AssetValue>,
+    C::AssetValue:
+        CheckedAdd<Output = C::AssetValue> + CheckedSub<Output = C::AssetValue> + SubAssign,
     C::Identifier: PartialEq,
 {
     type AssetMetadata = C::AssetMetadata;